Fixed more bug.
This engine really is something else. After 4 days I finally fixed a bug that wasn't even my fault.

So basically the engine delays the movement of Nodes that are a child of a PhysicsBody Node so that Tile positions are unreliable when the Grid is in motion. This caused SO many weird things to happen like random Tiles being destroyed, somehow tiles being off axis relative to the parent Grid and hundreds of "Ghost" Grids spawning.

To fix this I simply used dedicated "grid_position" properties for Tiles when performing calculations, instead of using the standard "position" property.