MMI Tile Rendering.
- Now using MultiMeshInstance2D (MMI) Nodes to render tiles, this is much more performant than rendering each sprite separately. Currently the texture is still being assigned per tile even though it is set to be invisible.

Currently tiles are assigned to an MMI through the tile's hashed type ID, however this makes all tiles of the same type use one single texture which   makes it impossible for variation in textures for same type tiles. I will rework the system so that tiles are assigned to MMIs based on the texture itself, not on the tile type id.

There is also now the issue of highlighting a mesh instance of a tile when it is hovered over, it is no longer as simple as before. So that will need reworked, for now tile highlighting doesn't work.

- Better player reach raycasting, can no longer reach farther diagonally.
- Added project license.